Code walk of implementation in java programming
The goal is to understand and analyze how the total cost and density of the network topology is dependent on number of nodes in the network connectivity of the designed network graph. With this a shortest path algorithm is implemented for constructing a network graph for achieving a graph with minimized transporting cost of the demand through the network.